Move search graph to utils
