+++ pyrobosim/pyrobosim/core/hallway.py
@@ -5,9 +5,9 @@ from shapely import intersects_xy
 from shapely.geometry import LineString, MultiLineString
 from shapely.plotting import patch_from_polygon
 
-from ..navigation.search_graph import Node
 from ..utils.pose import Pose, get_bearing_range
 from ..utils.polygon import inflate_polygon
+from ..utils.search_graph import Node
 
 
 class Hallway:
