+++ pyrobosim/pyrobosim/navigation/__init__.py
@@ -6,7 +6,6 @@ localization, mapping, path planning, and path following.
 
 # Utilities
 from .execution import *
-from .search_graph import *
 from .occupancy_grid import *
 from .trajectory import *
