+++ pyrobosim/pyrobosim/navigation/path_planner.py
@@ -47,10 +47,10 @@ class PathPlanner:
 
         :param start: Start pose or graph node.
         :type start: :class:`pyrobosim.utils.pose.Pose` /
-            :class:`pyrobosim.navigation.search_graph.Node`
+            :class:`pyrobosim.utils.search_graph.Node`
         :param goal: Goal pose or graph node.
         :type goal: :class:`pyrobosim.utils.pose.Pose` /
-            :class:`pyrobosim.navigation.search_graph.Node`
+            :class:`pyrobosim.utils.search_graph.Node`
         :return: Path from start to goal.
         :rtype: :class:`pyrobosim.utils.motion.Path`
         """
