+++ pyrobosim/pyrobosim/navigation/planner_base.py
@@ -38,10 +38,10 @@ class PathPlannerBase:
 
         :param start: Start pose or graph node.
         :type start: :class:`pyrobosim.utils.pose.Pose` /
-            :class:`pyrobosim.navigation.search_graph.Node`
+            :class:`pyrobosim.utils.search_graph.Node`
         :param goal: Goal pose or graph node.
         :type goal: :class:`pyrobosim.utils.pose.Pose` /
-            :class:`pyrobosim.navigation.search_graph.Node`
+            :class:`pyrobosim.utils.search_graph.Node`
         :return: Path from start to goal.
         :rtype: :class:`pyrobosim.utils.motion.Path`
         """
